requirements.txt
python2-paramiko
python-pymongo
python-IPy
python-tools
uwsgi
python-bottle
